Insert `-ms-filter` before `opacity`
Generally speaking, vendor-prefixed properties should always come before
the standard property they are prefixing. This allows browsers that
implement both the prefixed and the standard version to allow the
standard version to override.

Although `-ms-filter` is not a vendor prefix of `opacity`, it still
seems like it should appear before `opacity` after this transformation
has occurred. CSS-Tricks backs this up as well:

  https://css-tricks.com/almanac/properties/o/opacity/

Fixes #10